+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased][unreleased]
 
 - Add warning for top level 'type' property on an entity
+- Add 'Model.warnings' typings
 
 ## [1.2.2][] - 2021-05-17